+++ packages/facets/pnpm/src/features/detect.ts
@@ -51,6 +51,11 @@ export const detect: DetectFunction = async (ctx, dir) => {
     }
 
     if (detected) {
+      const pkgMan = pkgJson.getPackageManager()
+      if (pkgMan?.name === 'pnpm' && pkgMan?.version) {
+        detectResult.facet.scope.config.version = pkgMan.version
+      }
+
       return detectResult
     }
   }